Phase 6: unify cancellation-fee logic across manual and Flow payment paths
booking.py's manual/voucher cancel_booking previously applied zero fee
logic ("can always cancel free"), while payments_flow.py's Flow refund
path correctly applied the tiered cancellation-fee schedule. Both now
share app.services.cancellation.apply_cancellation(); manual cancellations
with money owed route into the same admin manual-refund queue Flow
provider failures already use.

Also fixes three bugs found while wiring this up: Booking.cancellation_fee
was typed Integer against a real numeric(12,2) column (silently truncated
cents), GET /admin/refunds/manual-queue 500'd on any non-empty result
(ManualQueueItem.booking_id didn't match Booking's actual field name), and
a client could claim cancelled_by_party="vendor" on their own booking to
dodge the client fee tiers entirely.

+++ backend/app/models/booking.py
@@ -28,7 +28,10 @@ class Booking(Base):
     updated_at = Column(DateTime, server_default=func.now(), onupdate=func.now())
 
     cancelled_at = Column(DateTime, nullable=True)
-    cancellation_fee = Column(Integer, nullable=True)
+    # DB column is numeric(12,2), matching refund_amount below -- was declared Integer here, which
+    # silently truncates cents on write (SQLAlchemy's Integer bind processor calls int() on the
+    # Decimal fee before sending it).
+    cancellation_fee = Column(Numeric(12, 2), nullable=True)
 
     # Added by the Phase 2 migration but not previously mapped on the ORM -- Phase 4 is the first
     # phase that actually reads/writes them.
